MM-48640 Move makeGetPostOptionBinding out of mattermost-redux
